Correction on the write function
The FOR statement has a variable  written increased by 1 in each loop. This cause a problem when you write more than one block  in the flash memory, which cause to lost bytes during this process. 

The variable written must change its value only when a new data is write in the cache or in the flash memory. (line 673)